+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.0
+  - KillGMSFont Module v1.0
+  - OMF extension v2022.7.6
+
 ### v0.1
 - Initial preparation
   - Universal module